refactor: split ganttGrid component
