+++ frontend/packages/design-system/src/components/gantt-view/ganttGrid.tsx
@@ -1,3 +1,6 @@
+/**
+ * External dependencies.
+ */
 import React, {
   useCallback,
   useEffect,
@@ -5,31 +8,18 @@ import React, {
   useRef,
   useState,
 } from "react";
-import { Plus } from "lucide-react";
-import {
-  ADD_PROJECT_ROW_HEIGHT,
-  CELL_HEIGHT,
-  HEADER_HEIGHT,
-} from "./constants";
+
+/**
+ * Internal dependencies.
+ */
+import { HEADER_HEIGHT } from "./constants";
 import { DeleteAllocationDialog } from "./deleteAllocationDialog";
-import { GanttMemberBar } from "./gantt-bar/memberBar";
-import { GanttProjectBar } from "./gantt-bar/projectBar";
-import {
-  RowAllocationOverlay,
-  type RowAllocationOverlayHandle,
-} from "./gantt-bar/rowAllocationOverlay";
-import { GanttMemberItem } from "./ganttMemberItem";
-import { GanttProjectItem } from "./ganttProjectItem";
+import { GanttMemberRows } from "./ganttMemberRows";
 import { createGanttStore, GanttContext, useGanttStore } from "./ganttStore";
-import type { Member } from "./ganttStore";
 import { GanttWeekHeader } from "./ganttWeekHeader";
 import { useContainerResize } from "./hooks/useContainerResize";
 import type { GanttGridProps } from "./types";
 import { mergeClassNames as cn } from "../../utils";
-
-type OverlayRefMap = React.RefObject<
-  Record<string, RowAllocationOverlayHandle | null>
->;
 
 const GanttGridInner: React.FC<{
   rootRef?: React.RefObject<HTMLDivElement | null>;
@@ -37,9 +27,7 @@ const GanttGridInner: React.FC<{
 }> = ({ rootRef, className }) => {
   const {
     members,
-    expandedRows,
     headerWidth,
-    daysPerWeek,
     columnWidth,
     weekendColumns,
     resizeHandleActive,
@@ -47,16 +35,11 @@ const GanttGridInner: React.FC<{
     startResize,
     columnCount,
     weeks,
-    hasRoleAccess,
-    onAddAllocation,
-    onEditAllocation,
     pendingDeleteEntry,
     clearPendingDeleteEntry,
   } = useGanttStore((s) => ({
     members: s.members,
-    expandedRows: s.expandedRows,
     headerWidth: s.headerWidth,
-    daysPerWeek: s.daysPerWeek,
     columnWidth: s.columnWidth,
     weekendColumns: s.weekendColumns,
     resizeHandleActive: s.resizeHandleActive,
@@ -64,9 +47,6 @@ const GanttGridInner: React.FC<{
     startResize: s.startResize,
     columnCount: s.columnCount,
     weeks: s.weeks,
-    hasRoleAccess: s.hasRoleAccess,
-    onAddAllocation: s.onAddAllocation,
-    onEditAllocation: s.onEditAllocation,
     pendingDeleteEntry: s.pendingDeleteEntry,
     clearPendingDeleteEntry: s.clearPendingDeleteEntry,
   }));
@@ -77,90 +57,6 @@ const GanttGridInner: React.FC<{
       startResize(e.clientX);
     },
     [startResize],
-  );
-
-  const canManageAllocations = hasRoleAccess && Boolean(onAddAllocation);
-  const canEditAllocations = hasRoleAccess && Boolean(onEditAllocation);
-  const memberRowOverlayRefs = useRef<
-    Record<string, RowAllocationOverlayHandle | null>
-  >({});
-  const projectRowOverlayRefs = useRef<
-    Record<string, RowAllocationOverlayHandle | null>
-  >({});
-
-  const setMemberRowOverlayRef = useCallback(
-    (rowKey: string, handle: RowAllocationOverlayHandle | null) => {
-      if (handle) {
-        memberRowOverlayRefs.current[rowKey] = handle;
-        return;
-      }
-
-      delete memberRowOverlayRefs.current[rowKey];
-    },
-    [],
-  );
-
-  const setProjectRowOverlayRef = useCallback(
-    (rowKey: string, handle: RowAllocationOverlayHandle | null) => {
-      if (handle) {
-        projectRowOverlayRefs.current[rowKey] = handle;
-        return;
-      }
-
-      delete projectRowOverlayRefs.current[rowKey];
-    },
-    [],
-  );
-
-  const getOverlayPointerDownHandler = useCallback(
-    (
-      overlayRefs: OverlayRefMap,
-      rowKey: string,
-      enabled: boolean,
-    ): React.PointerEventHandler<HTMLTableRowElement> | undefined => {
-      if (!enabled) {
-        return undefined;
-      }
-
-      return (event) => {
-        overlayRefs.current[rowKey]?.handleRowPointerDown(event);
-      };
-    },
-    [],
-  );
-
-  const getOverlayPointerMoveHandler = useCallback(
-    (
-      overlayRefs: OverlayRefMap,
-      rowKey: string,
-      enabled: boolean,
-    ): React.PointerEventHandler<HTMLTableRowElement> | undefined => {
-      if (!enabled) {
-        return undefined;
-      }
-
-      return (event) => {
-        overlayRefs.current[rowKey]?.handleRowPointerMove(event);
-      };
-    },
-    [],
-  );
-
-  const getOverlayPointerLeaveHandler = useCallback(
-    (
-      overlayRefs: OverlayRefMap,
-      rowKey: string,
-      enabled: boolean,
-    ): React.PointerEventHandler<HTMLTableRowElement> | undefined => {
-      if (!enabled) {
-        return undefined;
-      }
-
-      return () => {
-        overlayRefs.current[rowKey]?.clearHoveredSlot();
-      };
-    },
-    [],
   );
 
   return (
@@ -207,228 +103,9 @@ const GanttGridInner: React.FC<{
         </thead>
 
         <tbody>
-          {members.map((member: Member, rowIndex: number) => {
-            const isExpanded = expandedRows.has(rowIndex);
-            const animatedRowHeight = isExpanded ? CELL_HEIGHT : 0;
-            const addProjectRowHeight = isExpanded ? ADD_PROJECT_ROW_HEIGHT : 0;
-            const memberRowKey = `member-${rowIndex}`;
-
-            return (
-              <React.Fragment key={rowIndex}>
-                {/* Member row */}
-                <tr
-                  className="relative last:border-b border-outline-gray-1"
-                  onPointerDown={getOverlayPointerDownHandler(
-                    memberRowOverlayRefs,
-                    memberRowKey,
-                    canManageAllocations,
-                  )}
-                  onPointerMove={getOverlayPointerMoveHandler(
-                    memberRowOverlayRefs,
-                    memberRowKey,
-                    canManageAllocations,
-                  )}
-                  onPointerLeave={getOverlayPointerLeaveHandler(
-                    memberRowOverlayRefs,
-                    memberRowKey,
-                    canManageAllocations,
-                  )}
-                >
-                  <GanttMemberItem memberInd={rowIndex} />
-                  {weeks.map((_, i) => {
-                    return (
-                      <td
-                        key={i}
-                        colSpan={daysPerWeek}
-                        className={cn("border-r border-outline-gray-1")}
-                        style={{
-                          height: CELL_HEIGHT,
-                        }}
-                      />
-                    );
-                  })}
-                  <td
-                    aria-hidden="true"
-                    className="p-0 border-0 w-0 min-w-0 max-w-0"
-                    style={{ width: 0 }}
-                  >
-                    {member.memberAllocations.map((alloc, idx) => (
-                      <GanttMemberBar
-                        key={idx}
-                        allocation={alloc}
-                        memberInd={rowIndex}
-                      />
-                    ))}
-                    <RowAllocationOverlay
-                      ref={(handle) =>
-                        setMemberRowOverlayRef(memberRowKey, handle)
-                      }
-                      enabled={canManageAllocations}
-                      rowKey={memberRowKey}
-                      headerWidth={headerWidth}
-                      columnWidth={columnWidth}
-                      columnCount={columnCount}
-                      allocations={member.memberAllocations}
-                      createDraftBar={(left) => ({
-                        rowKey: memberRowKey,
-                        left,
-                        width: columnWidth,
-                        employeeId: member.id,
-                      })}
-                      onOpenAllocation={onAddAllocation}
-                    />
-                  </td>
-                </tr>
-
-                {/* Project child rows */}
-                {member.projects?.map((project, projectIndex) => {
-                  const projectRowKey = `project-${rowIndex}-${projectIndex}`;
-
-                  return (
-                    <tr
-                      key={`${rowIndex}-project-${projectIndex}`}
-                      className={cn("relative", {
-                        "pointer-events-none": !isExpanded,
-                      })}
-                      aria-hidden={!isExpanded}
-                      onPointerDown={getOverlayPointerDownHandler(
-                        projectRowOverlayRefs,
-                        projectRowKey,
-                        canManageAllocations && isExpanded,
-                      )}
-                      onPointerMove={getOverlayPointerMoveHandler(
-                        projectRowOverlayRefs,
-                        projectRowKey,
-                        canManageAllocations && isExpanded,
-                      )}
-                      onPointerLeave={getOverlayPointerLeaveHandler(
-                        projectRowOverlayRefs,
-                        projectRowKey,
-                        canManageAllocations && isExpanded,
-                      )}
-                    >
-                      <GanttProjectItem
-                        {...project}
-                        style={{
-                          height: animatedRowHeight,
-                          width: headerWidth,
-                          minWidth: headerWidth,
-                          borderBottomWidth: isExpanded ? undefined : 0,
-                          borderRightWidth: isExpanded ? undefined : 0,
-                        }}
-                      />
-                      {weeks.map((_, i) => {
-                        return (
-                          <td
-                            key={i}
-                            colSpan={daysPerWeek}
-                            className={cn(
-                              "overflow-hidden transition-[height] duration-200 ease-in-out",
-                              {
-                                "border-r border-outline-gray-1": isExpanded,
-                              },
-                            )}
-                            style={{
-                              height: animatedRowHeight,
-                            }}
-                          />
-                        );
-                      })}
-                      <td
-                        aria-hidden="true"
-                        className="p-0 border-0 w-0 min-w-0 max-w-0"
-                        style={{ width: 0 }}
-                      >
-                        {isExpanded &&
-                          project.allocations?.map((alloc, allocIndex) => {
-                            return (
-                              <GanttProjectBar
-                                key={allocIndex}
-                                allocation={alloc}
-                                resizable={canEditAllocations}
-                              />
-                            );
-                          })}
-                        <RowAllocationOverlay
-                          ref={(handle) =>
-                            setProjectRowOverlayRef(projectRowKey, handle)
-                          }
-                          enabled={canManageAllocations && isExpanded}
-                          rowKey={projectRowKey}
-                          headerWidth={headerWidth}
-                          columnWidth={columnWidth}
-                          columnCount={columnCount}
-                          allocations={project.allocations ?? []}
-                          createDraftBar={(left) => ({
-                            rowKey: projectRowKey,
-                            left,
-                            width: columnWidth,
-                            employeeId: member.id,
-                            projectId: project.id,
-                            projectName: project.name,
-                            customerName: project.client,
-                          })}
-                          onOpenAllocation={onAddAllocation}
-                        />
-                      </td>
-                    </tr>
-                  );
-                })}
-
-                {/* Add project row */}
-                {canManageAllocations && (
-                  <tr
-                    className={cn("relative", {
-                      "pointer-events-none": !isExpanded,
-                    })}
-                    aria-hidden={!isExpanded}
-                  >
-                    <th
-                      className="sticky left-0 z-10 bg-surface-white border-b border-r border-outline-gray-1 pl-8 pr-3 font-normal text-left align-middle flex items-center gap-2 w-full overflow-hidden transition-[height] duration-200 ease-in-out"
-                      style={{
-                        width: headerWidth,
-                        minWidth: headerWidth,
-                        height: addProjectRowHeight,
-                        borderBottomWidth: isExpanded ? undefined : 0,
-                        borderRightWidth: isExpanded ? undefined : 0,
-                      }}
-                    >
-                      <button
-                        type="button"
-                        onClick={() =>
-                          onAddAllocation?.({ employeeId: member.id })
-                        }
-                        className="w-full h-full flex items-center gap-2 text-base font-medium text-ink-gray-9 overflow-hidden"
-                      >
-                        <Plus size={16} className="shrink-0" />
-                        <span className="truncate">Add project</span>
-                      </button>
-                    </th>
-                    {weeks.map((_, i) => {
-                      return (
-                        <td
-                          key={i}
-                          colSpan={daysPerWeek}
-                          className={cn(
-                            "overflow-hidden transition-[height] duration-200 ease-in-out",
-                            {
-                              "border-r border-outline-gray-1": isExpanded,
-                            },
-                          )}
-                          style={{ height: addProjectRowHeight }}
-                        />
-                      );
-                    })}
-                    <td
-                      aria-hidden="true"
-                      className="p-0 border-0 w-0 min-w-0 max-w-0"
-                      style={{ width: 0 }}
-                    />
-                  </tr>
-                )}
-              </React.Fragment>
-            );
-          })}
+          {members.map((_, rowIndex) => (
+            <GanttMemberRows key={rowIndex} memberInd={rowIndex} />
+          ))}
         </tbody>
       </table>
